% Auto-generated: do not edit by hand
\name{dashPaperdragon}

\alias{dashPaperdragon}

\title{DashPaperdragon component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dashPaperdragon(id=NULL, imageSrc=NULL, label=NULL, value=NULL,
zoomLevel=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{imageSrc}{Character. the tile source for openseadrgon}

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{value}{Character. The value displayed in the input.}

\item{zoomLevel}{Numeric. zoomLevel of the current OSD Viewer}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

